requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
bson==0.5.10
certifi==2018.4.16
chardet==3.0.4
Click==7.0
cloudinary==1.19.1
cloudinary-cli==0.3.4
Django==2.2
dnspython==1.16.0
Flask==1.1.2
Flask-gzip==0.2
Flask-PyMongo==2.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.7
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
MarkupSafe==1.1.1
marshmallow==2.21.0
mock==3.0.5
motor==2.1.0
multidict==4.7.6
Pygments==2.5.2
pymodm==0.4.0
pymongo==3.10.1
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
requests==2.18.4
rfc3986==1.4.0
sanic==20.6.3
sanic-base-extension==0.1.1
sanic-mongodb-extension==0.4.0
six==1.11.0
sniffio==1.1.0
sqlparse==0.3.1
ujson==1.35
umongo==2.2.0
urllib3==1.22
uvloop==0.14.0
virtualenv==16.5.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
